First wait for WAB deployment before acquiring static hooks
The WAB deployment triggers the job to do plugin generation
for the endpoints.  The plugin generation registers a static
hook that delays the plugin generation until the process is
restored.  Before it was possible that the WABInstaller
checkpoint hook would get called after acquiring the static
hooks.  Once this happens no new static hooks are allowed to
be added to the CheckpointPhase class.

The issue is that some background thread may be still deploying
the system WABs and have yet to trigger the plugin generator.
This means the plugin generator may not have registered its hook
yet.  Ultimately this leads to the plugin generator not being
able to register its hook so it resorts to immediate plugin
generation on the checkpoint side.  This fix ensures that
all system WABs have been deployed before calling any static
checkpoint hooks.