For checkpoint add the -Djdk.nativeEC=false option
There are portability issues with nativeEC when doing
a checkpoint on AMD and restoring on Intel.  The reverse
doesn't seem to be an issue.  disabling nativeEC in the
JDK works around the issue.  This should not impact
performance to restore.

This is also a temporary setting while we work out the
real fix.