On restore always check for a delta in variable values
In cases where the configuration XML shows it is modified
the code previously did not check for a delta in the
variable values.  It only made a call to refreshConfiguration.
This did not cause the environment to be re-read for updated
values.  This resulted in cases where new config values did
not get read at all for restore.  The fix is to first
check for variablesChanged.  If there are any then simply call
variablesRefresh.  This also should result in the same processing
of refreshConfiguration with respect to reading updated config
XML, but it also will do the processing necessary for new ENV
values.