Disable bundle file closer for checkpoint
There is currently an issue with OpenJ9 JIT
for notify calls that cause an issue for checkpoint.

This disables the bundle file closer for checkpoint
only.  It remains to be seen that we need the closer
for checkpoint scenarios. Regardless, this is being
disabled for now until we get an OpenJ9 fix.