Make fattest.simplicity depend on Jackson bundle
This ensures that the Jackson bundle is included as a dependency when
building any FAT project which means it will be available to be copied
into the autoFVT zip.

Without this change the build works as long as the Jackson bundle is
included in the list of projects to build. However, if you try to
buildandrun a FAT test alone, gradle cannot find the Jackson project.