Comments explaining why TCKs require servlet
MP 6.0 depends on Jakarta EE Core which does not include servlet.
Therefore, we should be able to run all the MP 6.0 TCKs without servlet.

However, some TCKs do not work when switching the arquillian test
protocol from "servlet" to "rest" and so servlet is still required to
run the TCK.

Add comments to the relevant server.xmls pointing to why servlet is
still required.