Defer secure outbound chain startup
Since the SSL config can take a long time to turn up or never turn up at
all, we allow a dynamic reference to an object holding the SSL config.

If this never turns up, the outbound secure chain will never start up,
but the object will be available. This allows messaging to work until
someone tries to use the secure chain.