Change killIfRunning to killSilent
The UBI images do not have the ps command by default.
Instead of adding it to the image this change attempts
to kill the process silently to avoid having to invoke
the ps command to determin if the pid is still active
in order to avoid the error message when the pid no
longer exists.