Use SatisfyingConditionTarget to simplify process running condition
OSGi R8 DS spec added the satisfying condition target which
is an implicit reference to a condition.  By default all
components have an implicit reference to the "true" condition
which is always available from the framework.  The implicit
reference has the name "osgi.ds.satisfying.condition". This
name allows a component to specify a different target to
match another condition besides the "true" condition by
setting the component property named
"osgi.ds.satisfying.condition.target".

Furthermore the annotation SatisfyingConditionTarget annotation
was added to simplify this by only requiring the condition
matching filter to be specified. Here we are using the filter

SatisfyingConditionTarget("(osgi.condition.id=io.openliberty.process.running)")

That matches the "running" condition that is always available on a
checkpoint restored process or, in the case of a "normal" launch,
the condition is always available when the kernel is launched..

This allows the component to delay satisfied resolution until after
a checkpoint process is restored.