Control client filter creation in Telemetry
Previously, we provided the client filter class to JAX-RS and relied on
it to create instances as required.

However, JAX-RS was sometimes not using the right application context -
see #23985 - so I'm moving the creation of the filter instances into
Telemetry where we can delegate the creation to CDI which does get the
application context correct.