Close Telemetry span in Servlet listener
Use a ServletRequestListener to close the Telemetry Span, rather than
doing it in a JAX-RS response filter.

When calling async resource methods, request and response JAX-RS filters
can be called on different threads whereas the context scope must be set
and removed on the same thread.

Using a ServletRequestListener allows us to listen for when processing
is finished on the request thread, even if the response to the request
hasn't been sent yet.