Telemetry: don't set context for client
Don't set the current context in the client filter. There isn't any
need, the client span is only supposed to cover the actual request and
its details are propagated via the request headers. There's nothing else
that should run within the client span so there's nothing that should
need to have the client span set as the current context.

Never setting the current span means we don't have to worry about
closing the scope, though we do still have to worry about ending the
span. This isn't an issue at the moment, but potentially could be if the
response was handled on a different thread to the request.