Telemetry: document tests via JAX-RS as deliberate
Document that we want to keep the structure of the tests in
JaxRsEndpoints.

Although these tests would be clearer if re-written to use FATServlet,
keeping them in this style means that we have test coverage using the
JAX-RS and MP Rest Clients when called via JAX-RS as well as when called
via servlet.