Check that restore succeeded
When restoring from a checkpoint the restore can fail.
When this happens the server script will recover with
a normal launch.  Depending on the checkpoint testcase
this could go undetected and show the test passing.

But this would be testing against a non-restored server
which is not what the testcase would be trying to do.
This change forces the testcase to fail if we detect
that the server did not restore successfully.