Ensure ldapRegistry and federatedRegistry work for checkpoint
For InstantOn we delate the activation of service component
LdapAdapter until the restore side.  This is to avoid making
any connection to an LDAP server on checkpoint but also to
ensure the component gets the correct configuration on the
restore side without any extra churn.

The test case is a copy of the URAPIs_Federation_2LDAPsTest
test from com.ibm.ws.security.wim.adapter.ldap.fat FAT.
The test was modified to have the config for the LDAP
servers ports and addresses be configured on the restore
side through the use of environment variables.