Resolver: No extra autofeatures for resolveAsSet
Don't return any autofeatures which are unrelated to the set of features
requested when resolveAsSet is called, even if they would be satisfied
by a combination of installed and resolved features.

This is important because it's not clear how dependencies of any extra
selected autofeatures should be resolved, particularly if the
autofeature or its dependencies have tolerates dependencies.

In general, we can only resolve tolerates dependencies correctly when
we're resolving as a set and the feature with the dependency is included
in the set of features we're resolving.