The checkpoint (InstantOn) related message IDs need fixed.
When the checkpoint (InstantOn) messages were added there
was another PR adding messages around the same time for
Java-18 security manager changes.  This resulted in
more than one message using the same message ID.
Since InstantOn is still in beta it is acceptable
to change the message ID for it here to be correct.