Resolver: always prefer installed features
Fix a small bug where it was possible for KernelResolverRepository to
return a feature from the repository instead of one that was installed
if the repo feature was added to the repo first and had the same version
as the installed feature (which would usually be the case).

In practise, we would not hit this problem because we always add
installed features first.

Update unit test to show the problem.