No special autofeature treatment with resolveAsSet
If an autofeature name is passed to resolveAsSet(), don't attempt to
find and install any features which enable that autofeature. Instead
resolve it exactly the same way as the kernel resolver would, just
taking its actual dependencies.

The old behaviour was likely never used since we only do resolveAsSet
when the user has provided a server.xml and autofeatures wouldn't be
listed in the server.xml.

The old behaviour may also have caused problems if a public user feature
was enabled by a private feature. We would pass the private feature name
to the kernel resolver which would reject it.

Basic resolve is not affected by this change because
resolveFeaturesBasic already looks for features which enable an
autofeature when finding dependencies of requested features.