Do a resolveBundles call after installing all kernel bundles
This avoids having to resolve bundles while starting the kernel
bundles.  There is complications with resolving bundles in parallel
from the start-level.  While that should work, it is a complication
that we can easily avoid and this is better behavior anyway to
resolve all the kernel bundles in one call.