WIP: Fixes to use mvn wrapper
Use the standard .m2 directory if the user isn't configured to use
artifactory. Use standard mirror configuration in the settings.xml in
this case.

Keep the default values in maven-wrapper.properties, only overwrite them
if the user has configured artifactory.

Don't use Machine.execute because it merges arguments and then splits
them on whitespace. This requires our own handling of the process
output which we do asynchronously so that we can kill off the process if
it takes too long.

Split the maven output into lines as we read it to avoid repeatedly
splitting and merging it.