Grab exclusive lock to logging on checkpoint
This change is to avoid possible deadlocks in logging during checkpoint
when the JVM goes into single threaded mode.  Before going into
single-theaded mode an exclusive write lock is obtained by the thread
preparing to checkpoint.  This prepare thread will be the single running
thread once the JVM switches to single-threaded mode. This ensures that
other threads will not be in the middle of logging once we enter
single threaded mode.