Avoid deleting created dropins folder before checkpoint
If a dropins folder is created before checkpoint there are
cases where the folder could be in the middle of being processed
when a checkpoint happens.  Deleting such a directory that is
in the middle of being monitored (i.e. getting a file listing)
will prevent the process from being restored successfuly by
criu.

This fix avoids deleting such directories created before a
checkpoint.  This allows the process to be restored multiple times.