Repository resolver optimisations
There are several points in the resolution logic where we walk the
dependency tree because we want to take an action for every dependency
of a set of features. If we do this naively we end up walking over some
features many times because there are multiple ways to reach the feature
in the dependency tree from the starting set.

This is particularly a problem when we're searching for dependencies of
installed features. If many features are installed, and we want to find
any missing dependencies of all of them, walking the dependencies of
each feature results in a lot of duplicated work.

To avoid this, we add a new feature to FeatureTreeWalker to say we only
want to see a feature (and its children) once, no matter how many times
we encounter it in the dependency tree.

We then update the RepositoryResolver to use this option where
appropriate. Since it's the FeatureTreeWalker instance that remembers
which features have been seen, we also need to create one walker and
reuse it, rather than creating several in some places.