Add com.ibm.ws.cdi.2.0.jsf.jakarta to faces-4.0.
This bundle used to be added using the cdi-3.0/faces-3.0 auto feature.
Starting  with the faces-4.0 feature the cdi-4.0 feature is enabled automatically.
There is no need for an auto feature any longer and we can directly add the
com.ibm.ws.cdi.2.0.jsf.jakarta bundle to the faces-4.0 feature.