Repo resolver install lists performance
Improve the performance of the repository resolver creation of install
lists.

When the install list is long, finding and removing a feature from the
list is slow. Previously we were doing this for every feature we found
in the dependency tree.

Instead, map from the feature to the last position in the tree walk that
we found it. Then at the end we can sort the map entries by position.

Add some unit tests which do a resolve with a large number of features
so that we can more easily see if we've introduced a performance
regression.