Avoid reparsing repository JSON
Avoid reparsing the whole JSON when getAllAssets is called for a
zip-based or single-file repository.

Assume that the repository contents is unchanged as long as the size and
last modified date of the file are the same.

Parts of the repository code assume that querying for an asset twice
will return two different objects. To enable caching while still
returning a new object, add a copy constructor to all model objects.