Avoid distracting output on stderr after a checkpoint
On bash, when criu kills the server process, bash writes a message to standard error containing the line number and line of the script where java was launched. This is redirected to /dev/null with this change. NOTE this does not affect execution on bourne shell where the message is coming from the STDERR of the java process itself. The message is not as distracting in that case since it's simply the word Killed.