Improve performance of createInstallLists
The previous method of creating install lists required walking the whole
dependency tree, including walking the same feature multiple times if it
was reachable in the dependency tree via multiple routes.

However, this requires several orders of magnitude more visits than if
we only have to visit each feature once.

Instead, we now build up a map of the longest dependency chain from a
root feature to the feature in question. This is easily computed and
memoized by finding the longest dependency chain for all the features
that depend on it, and adding 1 to the longest. This approach only
requires visiting each feature once and is thus much faster.