Fix trailing comma in feature files
- With the trailing comma, it makes it look like `IBM-App-ForceRestart:
uninstall` and `install` are IBM API package names with no package type
specified.  Additionally applications are not restarted when the feature
is restarted because IBM-App-ForceRestart is not actually a property in
the feature file.  Luckily REST features depend on restfulWSClient and
other features which do have the IBM-App-ForceRestart on them so it may
not cause problems.