Attempt to fix BDS NONE with EJBs
Register EJB descriptors in the BDA, but don't give them to Weld. (Weld
will make beans of them if we do.)

Use the ones from the BDA when looking up the EJB Descriptor for a
managed object.

Make sure that InjectInjectionObjectFactory can also handle the EJB
Descriptor not being present in the bean manager.