Set content-length header to match size of updated content
The Content-length header was not being updated when the document path was changed. such that only paths of 7 characters could work.

remove flush from wrapper to prevent commit of content before the filter updates the response

Update test to not have a doc path value that is the same length of path as `/openapi`