Deal with Servlet 6 '/' behaviour in FileHandler
Under servlet 6, repeated '/' in the URI get collapsed to a single '/'.
This breaks some file paths passed in the URI to the file transfer
API. Use the same fix that was previously used for mbean operations
to repair the collapsed slashes, but tweak to deal with a leading
'/' on the path parameter.