Issue 25927: Fix JWK retrieval logic for JWKs missing 'use' entry
This change boils down to no longer considering `"use"` to be a key identifier. That entry denotes how the key is expected to be used, and can still be used to look for the appropriate key with the corresponding value if one is available. However if `"use"` is not in the JWK, the current logic cannot successfully re-load that JWK from the internal cache if our internal calls are looking for a specific `"use"`. This change should allow our logic to essentially fall back to loading a JWK without a `"use"` entry in the event we don't find one cached.

Resolves #25927