Jakarta REST updates
- Remove overlay files added in PR #24732 related to RESTEASY issue 3283
that aren't needed now that we use RESTEasy 6.2.3.  Leaving one file
from the PR since the fix for that file isn't available until 6.2.4.
When we move to that new version of RESTEasy the 3rd file can be
removed.
- Remove XML Binding internal API from restfulWS-4.0 feature since the
plan is to remove the APIs that had a dependency on XML Binding APIs.
This PR is changing the 4.0 feature to not do what is in 3.0 and 3.1
that was introduced in PR #24267.