Harden OpenAPI HTTPS Tests
Especially on Windows were certificate creation can be exceedingly slow, >30 seconds. We should wait for the server to state that the HTTPS endpoint is ready to accept traffic, before connecting the selenium driver.

It has been regularly seen that the cert creation on Windows will take over 30 seconds and in some scenarios were it only tests the HTTPS endpoint such as the OAuth Tests that cert creation will not even be completed by the time it has attempted to run the test. based on observations the timeout for the log checking is set to 60 seconds after server start, though cert creation does start before then.