Improve ProxySupportUtil performance
Due to ProxySupportUtil always going to fetch the bundle it has caused a performance degredation of ~7%. Instead of getting the context on every request, use appropiate mechanisms to map to the OpenAPIEndpointProvider.

For OpenAPI and MPOpenAPI 1.0, use a @Reference to obtain the Endpoint Provider in the ApplicationProcessor Component.

For MPOpenAPI 2.0 onwards use a ServiceTracker in the Servlet.

The Look up for the UI to inject the doc endpoint were already done by a service tracker, as such no changes are being made there.

Null checks in the ExtractUrl method already cover the possibility that the Provider is at the time `null`