Move SSL log Wait into central place
Instead of having the same wait in four places, put in the fattest LibertyServer definition so can be more easily reused. The method will throw a RuntimeException if we timeout after the default timeout period waiting for the SSL Listener to get ready. This is most likely to occur if the server takes >120 seconds to complete the generation of the certificates

For OAuth tests store and restore the base server config for the tests to stop potential bleeding of the Customized test case into the default causing the OpenAPI endpoints from not being in their expected location.

Add a wait for the Logout buttons appearance before executing the assertion.

Change Selenium waits to better match their example code for waits.

Replace hard-coding of OAuth2 redirect port