Reduce TCK maven request timeout
Reduce the request timeout (the time that maven is willing to wait for
the repository to send the next bit of data) to try to avoid hitting the
timeout for the whole project.

The default request timeout is 30 minutes and the whole test timeout is
one hour in LITE mode. We've occasionally seen one download from the
maven repository appear to hang, causing the project to timeout.