Include Smallrye RM API as separate bundle
Smallrye Reactive Messaging has it's own version of the Reactive
Messaging API which includes extensions.

Use their version at runtime, but ensure that we package the spec
version in our dev directory.

Create a new bundle for the Smallrye API, change the existing API bundle
to use the spec API and update the feature file to use the right one at
runtime.