Default Smallrye Reactive Messaging to Strict Mode
By default Smallrye RM is not MP Reactive Messaging spec compliant as it emit warnings if in a number of cases should emit a DeploymentException such as when an @Outgoing does not have a corresponding @Incoming.

Strict mode makes Smallrye RM spec compliant based on the tck results.

Add in default property enabled when mpReactiveMessaging-3.0 is enabled.