Update EE 11 features for wsSecurity and batchManagement features
- Public feature tolerates in xmlWS-4.0 and batch-2.1 cause the
wsSecurity and batchManagement features not to resolves when used with
EE 11 features.  This change updates to not use public tolerates, but
instead to use private features with tolerates to remove the need for
the features to include tolerates on EE 10 and EE 11 features in a way
that would make the features not work with previous EE feature levels
correctly.