Add Support for MP metrics
Previously the MP Metrics tests were ignored due to the fact there was no mp metrics API available and TestNG was the primary running of the test cases.

Removing testng now causes the metrics related tests to run. however the feature did not include support for metrics and it was not included within the class.

Add EmitterImpl copyright to show origin and changes made