Fix app startup failure for Reactive Messaging Fat tests
A number of the test packages contain multiple test servlets which all use unique channels that don't get injected into the mp config in other test causes leading to invalid graphs.

The change to strict mode, meant that what were previously warnings are now being reported as DeploymentExceptions in line with the specification.

Either add only the required classes or remove the bean that is not having channels configured for it