Remove hardcoding of port 8020 from OAuthFlow
Previous PR was a partial fix for Oauth Flow URL issue.

Remove host details from OAuth Flow URLs. The URLs can be relative to the application so this is a valid definition for the Urls.

Using the default Fat Test ports, when the new tab is created when clicking `Authorize` will go to the HTTP port causing an SSL error as OAuth flows can only be done via HTTPS. So by disabling HTTP port all redirects go to the https port.

fatTestPorts.xml is no included as on server definition restore the ordering can change meaning that it can appear 2nd in the server.xml and therefore provide the values. So fatTestCommon.xml is included with the new httpEndpoint definition, along with a cut down version of fatTestPorts