Rename RM common test library.
Due to build patterns relating to identifying test projects by their names. generally anything with `_fat` in the name is expected to produce a autoFAT.zip. As such the new internal_fat_common project which contains framework and basic classes for tests was being treated as a test project and therefore failing in the build.

However project names that match `*fat.common` can be recognized as test dependencies, but not to produce certain outputs expected from FAT projects.

Update references to the new common package to use ``.common` instead of `_common`