Issue 26225: Fix OIDC social client emitting raw NLS message key instead of message text
Updates the `RedirectionProcessor` class to have a local `Tr` variable that is tied to the NLS messages associated with its package. As the code was written, the `RedirectionProcessor` class accepted a `Tr` instance variable in its constructor and inadvertently used that object to look up an NLS message that wasn't in the associated message bundle. That caused the message to not be found, and the message key was emitted to the logs as-is. Now the local `Tr` variable will emit the correct error message.

Resolves #26225