Improve use of WSContextService in MP RSO 3.0
Make sure we're using the internal WSContextService singleton (which
does not require the concurrent feature and is not configurable by
users) and that we're submitting tasks to liberty's global thread pool
directly.

We may want to check if async is necessary here. If so, we may want to
consider allowing the user to configure a different context service.

Use the mutiny SPI ExecutionModel to instruct mutiny to use an executor
which calls WSContextService. This replaces the previous code which wrapped
one CompletionStage around another.