Update to use ReentrantLock instead of synchronized
- When running with virtual threads, thread cannot unmount if they are
in a synchronized block.  In order to work with virtual threads, the
stacks need to be updated to use ReentrantLock instead to allow for
unmounting.  The changed classes are ones that were identified as
needing to be updated due to threads not unmount when doing a Jakarta
RESTful Web Services work load using Jakarta EE 10 features when using
virtual threads.