Force interpretation of Swagger UI js as UTF-8
By not setting the charset on the Javascript it can potentially be interpreted as ISO-8859-1 causing `&nbsp;` to be seen as `Â ` (space included) due to the fact that they have the characters have matching byte code.

Setting the charset on the script in the html files forces the browser to view that file in the stated character set.

Apply to all html files for Swagger

Modify UI Basic Selenium tests to check the value of the text for a required parameter matches the expected UTF-8 value for it. if it contains the alternative it will fail