Fix Telemetry WithSpan + Agent on EE7
The existing solution for adding beans wasn't working on EE7. If there's
a beans.xml with bean-discovery-mode=none then all the beans returned by
SPIMetaData were ignored.

Previously the beans.xml was ignored because it used the Jakarta
namespace. This caused the agent tests to fail because the
WithSpanInterceptor was not disabled when the agent was detected.

When this was corrected, all the tests started failing because the beans
weren't registered at all.

The new approach removes the beans.xml and relies on an Extension to
veto the WithSpanInterceptor if the agent is detected.