Instrument injected Rest Clients on EE7 and EE8
Add a new hook point to MP Rest Client 1.x-2.x which allows us to
recieve a callback when a RestClientBuilder is created in order to
create an injected client.

Use this hook point to register the TelemetryClientFilter.

This isn't needed for MP Rest Client 3.x and newer because it already
calls our TelemetryRestClientBuilderListener for injected clients.