Update Swagger-ui Update Action
Fix checkout deprecation to support node16

Have the action commit to the current branch rather than opening a pull
request if the branch is not integration. This makes it easier to create
a branch with the required code changes and then use the action to
rebuild them in the same PR.

Update the action to only copy scss files from the swagger-ui source.