Ignore ISE while stopping the server
The tests do a stop server before updating the config. This stops the WAB io.openliberty.microprofile.openapi.2.0.internal.servlet.
But sometimes it throws ISE: BundleContext is no longer valid,  intermittently when ungetting the service because the service is chenged and the service tracker tries to unget the service and notices that the bundle context was closed and invalid.

Updated Equinox would no longer throw exceptions on invalid context when ungetService is called.
For reference: https://github.com/eclipse-equinox/equinox/issues/172