Update OpenAPI update workflow to include permissions
Currently, the OpenAPI workflow does not define its permissions. Add deny-all permissions to the workflow and add contents and pull-request permissions to the openapi job so if any new jobs are added they need to add their own permissions.

Update checkout and pull-request actions to latest to resolve new deprecations