Add option to cache data from directory repo
If we assume that the files on disk do not change, we can cache the data
from the repository instead of re-reading it.

This option should be used in all production scenarios to improve
performance.

Most of our tests want to write assets to the repository and then read
them back and so cannot use this option.