Add Beta guards to Liberty-kafka connector new options
fast.ack and context.service are new attributes for the liberty-kafka connector. However, this are not currently ready to GA and as such require beta guards on them.

If they are specified based on the RM specification they should be ignored as they outside of beta should not have any impact upon the operations

Update relate kafka tests to set the beta flag.