Update Upstream validation tests to use a method and a emitter
Make base validation tests run in LITE mode as

Move new KafkaServer stop to allow for the capturing on Kafka errors to the common fat library so that is can be reused in both sets of FATs. Add stop to Kakfa validation tests as initial test case

A change in Smallrye Reactive Messaging 4.12.0, now means that when emitters are validated on startup. If two or more emitter declarations use the same channel, then a single instance is injected into each reference to the channel.