Use unique group ids for custom context tests
Occasionally, during the running of these tests a reblancing of the consumer group used by all the connectors in the test might occur at a point during the test that means we get multiple reads of a ingoing channel by the connector. By assigning unique groups to the channels, we should no longer see any occasionally rebalancing.