Performance updates for application start
- Most of these are minor changes, but are on main line paths that can
have cumulative benefits.
- Add isDebugEnabled / isLoggable checks around trace calls.
- Update to use StringBuilder more efficiently by setting size and
reusing an explicit StringBuilder instance instead of a the compiler
making new ones each time.
- Switch to use String.replace(char,char) instead of
replace(String,String) for single character strings.
- Do not eagerly create and store hashText field in UtilsImpl Map
objects and just compute it each time since it is only used for trace.
- Update to only use add instead of contains and add when seeing if
something can be added to a Set.
- Do not get substring(1) in ZipFileContainer for no reason since
nestedContainerEntries is only used in that method.