More updates for Java security enablement with EE 11 features
- Don't use the ServerConfiguration to get the features since some tests
use invalid XML files and will result in parsing errors of the
server.xml
- Update the checking of java security errors to be based off of if the
server was started with Java security or not since now even if
isJava2SecurityEnabled returns true we may not have started it with Java
security due to other conditions like EE 11.  Some tests remove the
server.xml so then you get an error looking at the server.xml again to
see if it had EE 11 features.  This is resolved now as well.
- Update to handle whatever case someone uses for the feature names
since feature names are case insensitive.