Only check SAF if user's length is less than 8 characters
Currently `isValidUser` checks if user is valid all the time. However this shouldn't be the case because SAF only supports users that are less than 8 characters.
[https://www.ibm.com/docs/en/zos/2.4.0?topic=users-user-naming-conventions](https://www.ibm.com/docs/en/zos/2.4.0?topic=users-user-naming-conventions)

Checking for users that are longer than 8 characters causes this RACF exception and ffdc:

```
    [7/13/23 8:32:38:113 GMT] 0000003e id=00000000 com.ibm.ws.logging.internal.impl.IncidentImpl                I FFDC1015I: An FFDC Incident has been created: "com.ibm.ws.security.registry.RegistryException: CWWKS2910E: SAF service WAS_INTERNAL did not succeed. SAF return code 0xffffffff. RACF return code 0xffffffff. RACF reason code 0xffffffff. Internal error code 0x00000001. com.ibm.ws.security.wim.adapter.urbridge.URBridge 1008" at ffdc_23.07.13_08.32.38.0.log

```

In a single SAF UR environment, this can be acceptable since customers will not be able to create users longer than 8 characters without SAF Complaining. However in a VMM setup where LDAP and other UR are used that don't have this restriction, this can lead to the creation of hundreds if not thousands of ffdcs. This is because the VMM checks if a user exist in the Realm.

This MR will only check SAF if the user is less than 8 characters long.